Refactoring and switching from flag to cobra
